// src/api/generated/model/atsRefKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export type AtsRefKind = typeof AtsRefKind[keyof typeof AtsRefKind];


export const AtsRefKind = {
  greenhouse: 'greenhouse',
  lever: 'lever',
  workday: 'workday',
  personio: 'personio',
  ashby: 'ashby',
  smartrecruiters: 'smartrecruiters',
  workable: 'workable',
  recruitee: 'recruitee',
} as const;